Update readme.md
It was decided to leave both runs of the load stage. Because the biggest part of mapping of concepts from concept_manual is created automatically from ICD10 ancestors mapping. And we have to check if all concepts are mapped or mapped properly (if not - remap and run load_stage for the second time)